componentize the dropdown
