+++ client/src/App.tsx
@@ -32,7 +32,7 @@ import ToolsTab from "./components/ToolsTab";
 import ChatTab from "./components/ChatTab";
 import Sidebar from "./components/Sidebar";
 import Tabs from "./components/Tabs";
-import SettingsTab from "./components/SettingsTab";
+import SettingsTab from "./components/settings/SettingsTab";
 import ClientFormSection from "./components/ClientFormSection";
 import StarGitHubModal from "./components/StarGitHubModal";
